renamed routes-helpers.js to utils.js
